requirements.txt
# Core Telegram
python-telegram-bot>=20.0

# WebSocket
websockets>=11.0
websocket-client>=1.6.0

# Web Server (FastAPI)
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0

# Data Processing
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0

# Security
cryptography>=41.0.0
cryptography
fastapi
httpx
numpy
pydantic
python-dotenv
python-telegram-bot
telegram
uvicorn
websocket
websocket-client
websockets
